+++ packages/cli/src/commands/config/set.ts
@@ -1,13 +1,10 @@
-import {Command, Flags} from '@oclif/core';
+import {CliUx, Command, Flags} from '@oclif/core';
 import {Config} from '../../lib/config/config';
 import {AuthenticatedClient} from '../../lib/platform/authenticatedClient';
 import {
   IsAuthenticated,
   Preconditions,
 } from '../../lib/decorators/preconditions';
-import {PlatformEnvironment} from '../../lib/platform/environment';
-import {withEnvironment, withRegion} from '../../lib/flags/platformCommonFlags';
-import {Region} from '@coveord/platform-client';
 import {Trackable} from '../../lib/decorators/preconditions/trackable';
 import {InvalidCommandError} from '../../lib/errors/InvalidCommandError';
 import {ConfigRenderer} from '../../lib/config/configRenderer';
@@ -16,18 +13,19 @@ export default class Set extends Command {
   public static description = 'Modify the current configuration.';
 
   public static flags = {
-    ...withRegion(false),
-    ...withEnvironment(false),
+    environment: Flags.string({
+      char: 'e',
+      hidden: true,
+    }),
+    region: Flags.string({
+      char: 'r',
+      hidden: true,
+    }),
     organization: Flags.string({
       char: 'o',
       description:
         'The identifier of the organization inside which to perform operations. See <https://docs.coveo.com/en/1562/#organization-id-and-other-information>.',
       helpValue: 'myOrgID',
-    }),
-    analytics: Flags.string({
-      char: 'a',
-      options: ['y', 'n'],
-      description: 'Whether to enable analytics and telemetry tracking.',
     }),
   };
 
@@ -35,22 +33,18 @@ export default class Set extends Command {
   @Preconditions(IsAuthenticated())
   public async run() {
     const {flags} = await this.parse(Set);
+    if (flags.environment || flags.region) {
+      CliUx.ux.error(
+        'To connect to a different region or environment, use the `auth:login` command'
+      );
+    }
     if (Object.entries(flags).length === 0) {
       throw new InvalidCommandError('Command should contain at least 1 flag');
     }
     const cfg = new Config(this.config.configDir);
-    if (flags.environment) {
-      cfg.set('environment', flags.environment as PlatformEnvironment);
-    }
     if (flags.organization) {
       await this.verifyOrganization(flags.organization);
       cfg.set('organization', flags.organization);
-    }
-    if (flags.region) {
-      cfg.set('region', flags.region as Region);
-    }
-    if (flags.analytics) {
-      cfg.set('analyticsEnabled', flags.analytics === 'y');
     }
 
     ConfigRenderer.render(cfg);