+++ packages/cli/src/commands/org/resources/list.spec.ts
@@ -1,6 +1,6 @@
 jest.mock('../../../lib/config/config');
 jest.mock('../../../hooks/analytics/analytics');
-jest.mock('../../../hooks/prerun/prerun');
+
 jest.mock('../../../lib/platform/authenticatedClient');
 
 import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';