+++ packages/cli/src/lib/snapshot/snapshot.spec.ts
@@ -34,7 +34,6 @@ import {join} from 'path';
 import {SnapshotOperationTimeoutError} from '../errors';
 import {fancyIt} from '../../__test__/it';
 import {SnapshotReportStatus} from './reportPreviewer/reportPreviewerDataModels';
-import fancy from 'fancy-test';
 
 const mockedRetry = jest.mocked(retry);
 const mockedExpandedPreviewer = jest.mocked(ExpandedPreviewer, true);
@@ -51,8 +50,6 @@ const mockedExportSnapshot = jest.fn();
 const mockedApplySnapshot = jest.fn();
 const mockedDryRunSnapshot = jest.fn();
 const mockedGetClient = jest.fn();
-const mockedCreateSynchronizationPlan = jest.fn();
-const mockedApplySynchronizationPlan = jest.fn();
 const mockedGetSynchronizationPlan = jest.fn();
 //#endregion Mocks
 
@@ -89,8 +86,6 @@ describe('Snapshot', () => {
           get: mockedGetSnapshot,
           export: mockedExportSnapshot,
           apply: mockedApplySnapshot,
-          createSynchronizationPlan: mockedCreateSynchronizationPlan,
-          applySynchronizationPlan: mockedApplySynchronizationPlan,
           getSynchronizationPlan: mockedGetSynchronizationPlan,
         },
       })
@@ -295,9 +290,15 @@ describe('Snapshot', () => {
     });
 
     it.each([
-      [undefined, false],
-      [true, true],
-      [false, false],
+      [undefined, {deleteMissingResources: false}],
+      [
+        true,
+        {
+          deleteMissingResources: true,
+          deletionScope: 'ONLY_TYPES_FROM_SNAPSHOT',
+        },
+      ],
+      [false, {deleteMissingResources: false}],
     ])(
       'should request to apply the snapshot to the platform',
       async (validateParam, snapshotClientParam) => {
@@ -306,9 +307,10 @@ describe('Snapshot', () => {
 
         await snapshot.apply(validateParam);
 
-        expect(mockedApplySnapshot).toHaveBeenCalledWith(snapshotId, {
-          deleteMissingResources: snapshotClientParam,
-        });
+        expect(mockedApplySnapshot).toHaveBeenCalledWith(
+          snapshotId,
+          snapshotClientParam
+        );
 
         stderr.stop();
         stdout.stop();